Turn inspect example into otexplorer crate
I expect this project to expand, and I think it makes sense to have it
be a publishable crate.
