+++ otexplorer/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "otexplorer"
+version = "0.1.0"
+edition = "2021"
+license = "MIT/Apache-2.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+xflags = "0.2.4"
+read-fonts = { path = "../read-fonts"}
+font-types = { path = "../font-types"}
