+++ read-fonts/Cargo.toml
@@ -14,6 +14,3 @@ default = ["traversal"]
 [dependencies]
 font-types = { path = "../font-types" }
 bitflags = "1.3"
-
-[dev-dependencies]
-xflags = "0.2.4"
